Refactored uring-functions into a separate file
